Changed port

+++ src/commands/util/ping.js
@@ -7,17 +7,26 @@ module.exports = new Command({
     description: "Checks the bot's connection to the discord servers.",
 
     async run(message, args, bot) {
-        message.channel.send('🏓 Pinging....').then((msg) => {
+        if (message.author.id == 289896419291168775) {
             const embed = new MessageEmbed()
-                .setTitle('🏓 Pong!')
-                .setColor('PURPLE')
-                .setDescription(
-                    `Latency: ${Math.floor(
-                        msg.createdTimestamp - message.createdTimestamp
-                    )}ms\nAPI Latency: ${bot.ws.ping}ms`
-                );
-            msg.channel.send({ embeds: [embed] });
-            msg.delete();
-        });
+                .setTitle('No.')
+                .setImage('https://oskar.global/wp-content/uploads/2022/03/no.jpg')
+                .setColor('PURPLE');
+            message.channel.send({ embeds: [embed] });
+            return;
+        } else {
+            message.channel.send('🏓 Pinging....').then((msg) => {
+                const embed = new MessageEmbed()
+                    .setTitle('🏓 Pong!')
+                    .setColor('PURPLE')
+                    .setDescription(
+                        `Latency: ${Math.floor(
+                            msg.createdTimestamp - message.createdTimestamp
+                        )}ms\nAPI Latency: ${bot.ws.ping}ms`
+                    );
+                msg.channel.send({ embeds: [embed] });
+                msg.delete();
+            });
+        }
     },
 });
